requirements.txt
streamlit>=1.28.0
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
Pillow>=10.1.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.15.0
requests>=2.31.0
python-dotenv>=1.0.0